# src/routers/organisation_v1_router.py
from fastapi import APIRouter, Depends, Query, Body, Path
from src.services.organisation_v1_service import OrganisationService
from src.dependencies import get_config
from src.models.organisation_v1_schemas import Date, PersonsByPortfolioRquest
from src.services.core_opengin_service import OpenGINService

router = APIRouter(prefix="/v1/organisation", tags=["Organisation"])

def get_organisation_service(config: dict = Depends(get_config)):
    opengin_service = OpenGINService(config=config)
    return OrganisationService(config, opengin_service)

@router.post('/active-portfolio-list', summary="Get active portfolio list.", description="Returns a list of portfolios under a given president and a given date.")
async def active_portfolio_list(
    presidentId: str = Query(..., description="ID of the president"),
    body: Date = Body(...),
    service: OrganisationService = Depends(get_organisation_service)
    ):
    service_response = await service.active_portfolio_list(presidentId, body.date)
    return service_response

@router.post('/departments-by-portfolio/{portfolio_id}', summary="Get active departments for a portfolio.", description="Returns a list of departments under a given portfolio and a given date.")
async def departments_by_portfolio(
    portfolio_id: str = Path(..., description="ID of the portotfolio"),
    body: Date = Body(...),
    service: OrganisationService = Depends(get_organisation_service)
):  
    service_response = await service.departments_by_portfolio(portfolio_id=portfolio_id, selected_date=body.date)
    return service_response

@router.post('/persons-by-portfolio/{portfolio_id}', summary="Get active persons for a portfolio.", description="Returns a list of persons under a given portfolio and a given date.")
async def persons_by_portfolio(
    portfolio_id: str = Path(..., description="ID of the portfolio"),
    body: PersonsByPortfolioRquest = Body(...),
    service: OrganisationService = Depends(get_organisation_service)
):
    service_response = await service.persons_by_portfolio(portfolio_id=portfolio_id, president_id=body.president_id, selected_date=body.activeDate.date)
    return service_response

@router.post('/prime-minister')
async def prime_minister(
    active_date: Date = Body(...),
    service: OrganisationService = Depends(get_organisation_service)
):
    service_response = await service.prime_minister(selected_date=active_date.date)
    return service_response